[curielogger] Initialize templates, ILM, and index
Move the initialization steps from bash to curielogger directly. This will give us
more control on how, when, and why these steps are called. We will also be able to
improve the logic, and extend it (add versioning, for example).

As part of this change, I have moved the Elasticsearch implementation to its own file,
and I will be doing this for every output soon.

related to #99

Signed-off-by: Flavio Percoco <flavio@reblaze.com>